feat: add Cypress tests for Content Model, move intercepts into single function
