+++ cypress/e2e/contentLibrary/panel/objectPanel.cy.ts
@@ -1,10 +1,10 @@
-import { configurePanelSkylarkIntercepts } from "./__utils__/test-utils";
+import { configureSkylarkIntercepts } from "../../../support/utils/handlers";
 
 describe("Content Library - Object Panel", () => {
   beforeEach(() => {
     cy.login();
 
-    configurePanelSkylarkIntercepts();
+    configureSkylarkIntercepts();
 
     cy.visit("/");
     cy.wait("@introspectionQuery");
