+++ cypress/e2e/contentLibrary/panel/tabs/imagery.cy.ts
@@ -1,10 +1,10 @@
-import { configurePanelSkylarkIntercepts } from "../__utils__/test-utils";
+import { configureSkylarkIntercepts } from "../../../../support/utils/handlers";
 
 describe("Content Library - Object Panel - Imagery tab", () => {
   beforeEach(() => {
     cy.login();
 
-    configurePanelSkylarkIntercepts();
+    configureSkylarkIntercepts();
 
     cy.visit("/");
     cy.wait("@introspectionQuery");
