+++ cypress/e2e/contentLibrary/panel/tabs/metadata.cy.ts
@@ -1,11 +1,11 @@
 import { hasOperationName } from "../../../../support/utils/graphqlTestUtils";
-import { configurePanelSkylarkIntercepts } from "../__utils__/test-utils";
+import { configureSkylarkIntercepts } from "../../../../support/utils/handlers";
 
 describe("Content Library - Object Panel - Metadata tab", () => {
   beforeEach(() => {
     cy.login();
 
-    configurePanelSkylarkIntercepts();
+    configureSkylarkIntercepts();
 
     cy.visit("/");
     cy.wait("@introspectionQuery");
